Refactor GraphQL query location
