+++ pages/entries.js
@@ -1,11 +1,26 @@
 import { useState, useRef } from 'react';
-import { format } from 'date-fns'
 import DayPicker, { DateUtils } from 'react-day-picker';
-import DayPickerInput from "react-day-picker/DayPickerInput";
 import 'react-day-picker/lib/style.css';
+import gql from "graphql-tag";
+import { format, startOfDay, endOfDay } from 'date-fns'
+import { useQuery } from '@apollo/react-hooks';
+import { useStoreState } from 'easy-peasy';
+import { zonedTimeToUtc } from 'date-fns-tz';
 
 import withLayout from '../components/Layout';
 import EntriesList from '../components/EntriesList';
+
+const GET_ENTRIES = gql`
+  query UserEntries($userID: ID!, $startDate: String, $endDate: String) {
+    entriesByUserID(userID: $userID, startDate: $startDate, endDate: $endDate) {
+      id
+      wordCount
+      createdAt
+    }
+  }
+`;
+
+const timezoneOffset = new Date().getTimezoneOffset();
 
 const formatFriendly = date => format(new Date(date), 'MMMM d, yyyy');
 
@@ -14,9 +29,25 @@ const Entries = () => {
   const [endDate, setEndDate] = useState(null);
   const [enteredToDate, setEnteredToDate] = useState(null);
 
+  // GraphQL Query
+  const { uid: userID } = useStoreState(state => state.user).firebaseData;
+  const utcStartDate = startDate ? zonedTimeToUtc(startOfDay(startDate), timezoneOffset) : null;
+  const utcEndDate = endDate ? zonedTimeToUtc(endOfDay(endDate), timezoneOffset) : null;
+  const { loading, error, data } = useQuery(GET_ENTRIES, {
+    variables: {
+      userID,
+      startDate: utcStartDate,
+      endDate: utcEndDate,
+    },
+  });
+
   // Date Picker config
   let datePicker = useRef();
-  const modifiers = { start: startDate, end: enteredToDate };
+  const modifiers = {
+    start: startDate,
+    end: enteredToDate,
+    entries: [],
+  };
   const disabledDays = { before: startDate };
   const selectedDays = [startDate, { from: startDate, to: enteredToDate }];
 
@@ -101,7 +132,7 @@ const Entries = () => {
         </div>
       </div>
 
-      <EntriesList startDate={startDate} endDate={endDate} />
+      <EntriesList data={data} loading={loading} error={error} />
     </div>
   );
 };
